sort when casting set to list
Currently, when printing or saving designs, the block and trial factors will be differently ordered for every Python session. This is because sets don't have an order and when being casted to lists the assigned order seems to be random. I think we should explicitly order those factor names, such that the resulting design CSV file is better readable.